Added fish config

+++ linkfiles.sh
@@ -1,11 +1,17 @@
 #!/bin/sh
 
 for file in $(ls); do
-    [ -e ~/.$file ] &&
-	echo "Warning: ~/.$file exists, backed up to ~/.${file}.bkp" &&
-	rm -rf ~/.${file}.bkp &&
-        mv ~/.$file ~/.${file}.bkp
-    [ "$file" = "LICENSE" ] || [ "$file" = "linkfiles.sh" ] &&
-	continue
-    ln -s dotfiles/$file ~/.$file
+  if [ "$file" = "LICENSE" ] || [ "$file" = "linkfiles.sh" ] || [ "$file" = "config" ]; then
+    continue
+  fi
+  if [ -e ~/.$file ]; then
+	  echo "Warning: ~/.$file exists, backed up to ~/.${file}.bkp"
+	  rm -rf ~/.${file}.bkp
+    mv ~/.$file ~/.${file}.bkp
+  fi
+  ln -s dotfiles/$file ~/.$file
+done
+
+for file in $(ls config); do
+  ln -s dotfiles/config/$file ~/.config/$file
 done
